Move prediction files into dedicated subdirectory
