+++ src/python/spectrseqtools/prediction/composition_inference.py
@@ -6,7 +6,7 @@ import polars as pl
 import numpy as np
 
 from spectrseqtools.masses import NUCLEOTIDE_DF, UNMODIFIED_BASES
-from spectrseqtools.traceback_matrix import CompositionInferrer
+from spectrseqtools.prediction.traceback_matrix import CompositionInferrer
 
 
 @dataclass
