+++ src/python/spectrseqtools/prediction/skeleton_building.py
@@ -10,9 +10,9 @@ from spectrseqtools.common import (
     calculate_error_threshold,
     calculate_compositions,
 )
-from spectrseqtools.fragment_classification import MAX_VARIANCE
-from spectrseqtools.sequence_inference import LinearProgramInstance
-from spectrseqtools.traceback_matrix import (
+from spectrseqtools.prediction.fragment_classification import MAX_VARIANCE
+from spectrseqtools.prediction.sequence_inference import LinearProgramInstance
+from spectrseqtools.prediction.traceback_matrix import (
     CompositionInferrer,
     compute_sequence_length_bound,
 )
